Added a "Share your Plan" button
Added a "Share your Plan" button at the bottom of the plan chart. Now you can share your plan settings link with one click. This is convenient for training and teamwork. It is easy and simple to have the same settings for the student and the instructor. The link can be shared via WhatsApp, Telegram, LinkedIn, FB.

Updated Android SRC.