Fixed an inaccuracy when there are no decompression mixes in the plan
Fixed an inaccuracy when there are no decompression mixes in the plan and the plan has only one decompression stop that raises you immediately by 3 meters (10 feet), but the glider settings for the last stop are 6 meters (20 feet). Now it is processed correctly and when set to 6 meters (20 feet) it will raise you 6 meters (20 feet). Thanks for the report, Divert hhassey.