FIXED. In the graph display, temperature and immersion depth compensation was added twice
In the graph display, temperature and immersion depth compensation was added twice, which caused incorrect display of the PPO2, PPN2, and PPHe graphs in hundredths. The duplication has been removed, and now when rounded to tenths, instead of the incorrect 1.7, it will be 1.6, as it should be. The table and graph are displayed with an accuracy of hundredths, and it will be 1.68.